It now technically works!!
Block to Item works fully and can be used as normal.
Block to Block almost works fully, explanation below.

Some small additions that will be added soon are:
- A 'jitter' to the mouse cursor move slightly so that the game recognizes when we change block
- 'Centering' the mouse position after exiting a block/item edit mode (the next block is usually here, so we can then just click again)

Now block still needs a bit of work, since the game crashes somewhat constantly when using any type of conversion tool
we have to take this into acount, we therefore have to implement a check for if a 'block'/'item' already exists properly.
This will involve checking if the file exists AND checking if it's a _customblock (this then needs to be compared to a normal folder).
Though I still haven't decided on how to do this xdd